采用 index iterator 虽然效率上可能慢了一点点，但是更加直观，便于排错

+++ core/types.h
@@ -11,7 +11,7 @@
 #include "movedirection.h"
 
 typedef short type_value;
-typedef short type_size;
+typedef int type_size;
 
 typedef Basic_Pos<type_size> Pos;
